Backend/requirements.txt
opencv-python==4.8.1.78
PyYAML==6.0.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
ultralytics==8.0.203
numpy==1.26.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
